Test: REFACTOR -> change expected condition in one case
For short push when push is stable and short it's not possible to not call function that should be called on release so the timer need to be incremented.
It also make no sense to debounce the release state after pushing while debouncig will start when it will detect anoter push state... therefore some test has changed the expected value from 10 to 11.